IBM UrbanCode Standards Update
- Major restructuring of files and folders
- Added build.xml (Can now build with command: 'ant')
- Added ivy.xml for dependency management
- All scripts now use groovy-plugin-utils
- Updated all step names
