+++ .gitignore
@@ -1,0 +1,16 @@
+# ignore all dotfiles by except git specific ones
+.*
+!.gitignore
+!.git*
+!.travis.yml
+
+# ignore all temporary files and ahp specific files
+*.bak
+*.bom
+*.swp
+*~
+
+# ignore compilation related directories in the root of the project
+/releases
+/build
+/lib
