--- classes/com/urbancode/air/XTrustManagerFactory.groovy
@@ -1,58 +1,0 @@
-/*
-* Licensed Materials - Property of IBM Corp.
-* IBM UrbanCode Build
-* IBM UrbanCode Deploy
-* IBM UrbanCode Release
-* IBM AnthillPro
-* (c) Copyright IBM Corporation 2002, 2014. All Rights Reserved.
-*
-* U.S. Government Users Restricted Rights - Use, duplication or disclosure restricted by
-* GSA ADP Schedule Contract with IBM Corp.
-*/
-package com.urbancode.air
-
-import java.security.InvalidAlgorithmParameterException;
-import java.security.KeyStore;
-import java.util.ArrayList;
-import java.util.List;
-
-import javax.net.ssl.ManagerFactoryParameters;
-import javax.net.ssl.TrustManager;
-import javax.net.ssl.TrustManagerFactorySpi;
-
-public class XTrustManagerFactory extends TrustManagerFactorySpi {
-
-    //**********************************************************************************************
-    // CLASS
-    //**********************************************************************************************
-
-    //**********************************************************************************************
-    // INSTANCE
-    //**********************************************************************************************
-
-    //------------------------------------------------------------------------------------------
-    public XTrustManagerFactory() {
-    }
-
-    //------------------------------------------------------------------------------------------
-    @Override
-    protected void engineInit(KeyStore keystore) {
-    }
-
-    //------------------------------------------------------------------------------------------
-    @Override
-    protected void engineInit(ManagerFactoryParameters managerParams)
-    throws InvalidAlgorithmParameterException {
-        throw new InvalidAlgorithmParameterException(
-                "XTrustManagerFactory cannot be initialized using ManagerFactoryParameters");
-    }
-
-    //------------------------------------------------------------------------------------------
-    @Override
-    protected TrustManager[] engineGetTrustManagers() {
-        List<TrustManager> managerList = new ArrayList<TrustManager>();
-        managerList.add(new OpenX509TrustManager());
-
-        return managerList.toArray(new TrustManager[managerList.size()]);
-    }
-}
