+++ ivy.xml
@@ -1,0 +1,31 @@
+<ivy-module version="2.0">
+<!--
+    © Copyright IBM Corporation 2017.
+    This is licensed under the following license.
+    The Eclipse Public 1.0 License (http://www.eclipse.org/legal/epl-v10.html)
+    U.S. Government Users Restricted Rights:  Use, duplication or disclosure restricted by GSA ADP Schedule Contract with IBM Corp.
+-->
+    <info organisation="com.ibm.urbancode" module="dependencies"/>
+    <configurations>
+        <conf name="default" description="Jars used during runtime"/>
+        <conf name="build" description="Jars used during build"/>
+        <conf name="test" description="Jars used during testing"/>
+    </configurations>
+    <dependencies>
+        <!-- Runtime Dependencies -->
+        <!-- http://search.maven.org/#search%7Cga%7C1%7Cg%3A%22com.ibm.urbancode.plugins%22 -->
+        <dependency org="com.ibm.urbancode.plugins" name="groovy-plugin-utils" rev="1.2" transitive="false" conf="default->default"/>
+
+        <!-- Build Dependencies -->
+        <!-- http://search.maven.org/#search%7Cga%7C1%7Cg%3A%22com.ibm.urbancode.plugins%22 -->
+        <dependency org="com.ibm.urbancode.plugins" name="groovy-plugin-utils" rev="1.2" transitive="true" conf="build->default"/>
+
+        <!-- Test Dependencies -->
+        <!-- http://search.maven.org/#artifactdetails%7Cjunit%7Cjunit%7C4.11%7Cjar -->
+        <dependency org="junit" name="junit" rev="4.11" transitive="false"  conf="test->default"/>
+        <!-- http://search.maven.org/#artifactdetails%7Corg.hamcrest%7Chamcrest-core%7C1.3%7Cjar -->
+        <dependency org="org.hamcrest" name="hamcrest-core" rev="1.3" transitive="false" conf="test->default"/>
+        <!-- https://mvnrepository.com/artifact/org.codehaus.groovy/groovy-all -->
+        <dependency org="org.codehaus.groovy" name="groovy-all" rev="1.8.8" transitive="false" conf="build->default"/>
+    </dependencies>
+</ivy-module>
