--- service.groovy
@@ -1,57 +1,0 @@
-/*
-* Licensed Materials - Property of IBM Corp.
-* IBM UrbanCode Build
-* IBM UrbanCode Deploy
-* IBM UrbanCode Release
-* IBM AnthillPro
-* (c) Copyright IBM Corporation 2002, 2016. All Rights Reserved.
-*
-* U.S. Government Users Restricted Rights - Use, duplication or disclosure restricted by
-* GSA ADP Schedule Contract with IBM Corp.
-*/
-import com.urbancode.air.AirPluginTool;
-
-def apTool = new AirPluginTool(this.args[0], this.args[1]) //assuming that args[0] is input props file and args[1] is output props file
-
-def props = apTool.getStepProperties();
-
-final def isWindows = apTool.isWindows;
-
-def name = props["name"];
-def enabled = props["enabled"];
-def state = props["state"];
-
-def now = new Date();
-def time = now.getTime();
-def filename="tmp"+time+".yml";
-
-File file = new File(filename);
-file.write "- hosts: localhost\n";
-file << "  any_errors_fatal: true\n";
-file << "  tasks:\n";
-
-file << "  - name: service\n";
-file << "    service:\n";
-if (name) {
-file << "      name: "+name+"\n";
-}
-if (enabled) {
-file << "      enabled: "+enabled+"\n";
-}
-if (state) {
-file << "      state: "+state+"\n";
-}
-
-def command="ansible-playbook "+filename+ " -vvvv";
-println("command is "+command)
-def exec =command.execute();
-exec.waitFor();
-
-def error = new StringBuffer()
-exec.consumeProcessErrorStream(error)
-
-exec.text.eachLine {println it}
-println(error);
-def exitValue=exec.exitValue();
-if(!exitValue)
-	System.exit(exitValue);
